+++ lib/storage/src/content_manager/toc/collection_meta_ops.rs
@@ -122,26 +122,7 @@ impl TableOfContent {
             }
             #[cfg(feature = "staging")]
             CollectionMetaOperations::TestSlowDown(test_slow_down) => {
-                // Execute on this peer if: peer_id is None (all peers) OR peer_id matches this peer
-                let should_execute = test_slow_down
-                    .peer_id
-                    .is_none_or(|peer_id| peer_id == self.this_peer_id);
-
-                if should_execute {
-                    log::debug!(
-                        "TestSlowDown: sleeping for {}ms on peer {}",
-                        test_slow_down.duration_ms,
-                        self.this_peer_id
-                    );
-                    tokio::time::sleep(std::time::Duration::from_millis(
-                        test_slow_down.duration_ms,
-                    ))
-                    .await;
-                    log::debug!(
-                        "TestSlowDown: finished sleeping on peer {}",
-                        self.this_peer_id
-                    );
-                }
+                test_slow_down.execute(self.this_peer_id).await;
                 Ok(true)
             }
         }
